Feature request #25: Refactored Guice bindings for database ConnectionFactory. From now on, particular ConnectionFactory implementation is injected basis on annotation.
Because DatabaseJNDIPlugin can work only with JNDIConnectionFactory and DatabasePlugin can work only with StandaloneConnectionPool, particular implementation of ConnectionFactory is injected to these classes basis on @Named annotation value (jndi/standalone). It allows to overcome problems with overriding module bindings in the JNDI-aware environment, as it was done in the EnterpriseModule (removed with this commit).

git-svn-id: https://svn.code.sf.net/p/web-harvest/code/branches/release-2.1@560 a57e4815-81e3-4922-91ae-0d9c2f5eb974