Feature request #32: Removed AbstractProcessor/BodyProcessor #debug method. In all cases when #debug was called, we had hidden (quite expensive) cost of parameters construction.
In such case, it is better to first check if logger has desired level enabled. Of course it will incur twice the cost of evaluating whether logger level is enabled or not, but on the other hand, our logged parameters will be only computed in case of logger level enabled. It will reduce overhead significantly, especially because AbstractProcessor#run method is called for almost every single line of Scraper's configuration.

git-svn-id: https://svn.code.sf.net/p/web-harvest/code/branches/release-2.1@570 a57e4815-81e3-4922-91ae-0d9c2f5eb974