Bug #56 Reverted changes in JSRScriptAdatper to the previous revision.
ScriptEngines are no longer cached globally, but per thread. It allows JSRScriptEngineAdapter to copy bindings from scraper's DynamicScopeContext directly into ScriptEngine (one per thread). It is important to bear in mind that this implementation of JSRScriptEngineAdapter is not thread safe. We had to do some trade-off in order to gain on performance - currently creation of brand new javax.script.ScriptContext instances each time script is executed is quite expensive. It is hard to observe it on simple scraper configurations, but in case of more complex cases, web harvest with the previous adapter implementation worked approximately 4 times faster.

git-svn-id: https://svn.code.sf.net/p/web-harvest/code/branches/release-2.1@629 a57e4815-81e3-4922-91ae-0d9c2f5eb974