Bug #59: Modified StandaloneConnectionPool to use its .class classloader instead of system classloader when constructing dynamic proxy for ConnectionProxy interface.
When web harvest was embedded in web application running on tomcat we had "IllegalArgumentException: interface org.webharvest.runtime.database.ConnectionProxy is not visible from class loader". The reason is the hierarchy of Tomcat classloaders - system class loader which is a parent for common class loader, which in turn is parent for all web applications classloaders does not contain web applications classes (only bin/bootstrap.jar, bin/tomcat-juli.jar and bin/commons-daemon.jar). We have to use particular application class loader having access to all application classes, including ConnectionProxy (this is not the case in standalone applications).

git-svn-id: https://svn.code.sf.net/p/web-harvest/code/branches/release-2.1@632 a57e4815-81e3-4922-91ae-0d9c2f5eb974